Move generic provider creation functions into core.py
