+++ src/providers/core.py
@@ -1,18 +1,30 @@
+import copy
+from concurrent.futures import ThreadPoolExecutor
+from threading import Lock
 from typing import List, Tuple
 
+from app.provider.enum import ProviderStatus, ProviderType
 from app.provider.schemas_extended import (
     BlockStorageServiceCreateExtended,
     ComputeServiceCreateExtended,
     IdentityProviderCreateExtended,
     IdentityServiceCreate,
     NetworkServiceCreateExtended,
+    ProjectCreate,
+    ProviderCreateExtended,
     RegionCreateExtended,
     SLACreateExtended,
     UserGroupCreateExtended,
 )
 
+from src.logger import logger
 from src.models.identity_provider import SLA, Issuer, UserGroup
-from src.models.provider import PerRegionProps, Project, TrustedIDP
+from src.models.provider import PerRegionProps, Project, Provider, TrustedIDP
+from src.providers.openstack import get_data_from_openstack
+
+projects_lock = Lock()
+region_lock = Lock()
+issuer_lock = Lock()
 
 
 def filter_projects_on_compute_resources(
@@ -185,3 +197,134 @@ def update_region_network_services(
             break
     else:
         current_services.append(new_service)
+
+
+def get_project_resources(
+    *,
+    provider_conf: Provider,
+    project_conf: Project,
+    issuers: List[Issuer],
+    out_region: RegionCreateExtended,
+    out_projects: List[ProjectCreate],
+    out_issuers: List[IdentityProviderCreateExtended],
+) -> None:
+    # Find region props matching current region.
+    region_props = next(
+        filter(
+            lambda x: x.region_name == out_region.name,
+            project_conf.per_region_props,
+        ),
+        None,
+    )
+    proj_conf = get_project_conf_params(
+        project_conf=project_conf, region_props=region_props
+    )
+
+    try:
+        identity_provider, token = get_identity_provider_info_for_project(
+            issuers=issuers,
+            trusted_issuers=provider_conf.identity_providers,
+            project=proj_conf,
+        )
+    except ValueError as e:
+        logger.error(e)
+        logger.error(f"Skipping project {proj_conf.id}.")
+        return None
+
+    if provider_conf.type == ProviderType.OS.value:
+        resp = get_data_from_openstack(
+            provider_conf=provider_conf,
+            project_conf=proj_conf,
+            identity_provider=identity_provider,
+            token=token,
+            region_name=out_region.name,
+        )
+    if not resp:
+        return None
+
+    (
+        project,
+        block_storage_service,
+        compute_service,
+        identity_service,
+        network_service,
+    ) = resp
+    with region_lock:
+        update_region_block_storage_services(
+            current_services=out_region.block_storage_services,
+            new_service=block_storage_service,
+        )
+        update_region_compute_services(
+            current_services=out_region.compute_services, new_service=compute_service
+        )
+        update_region_identity_services(
+            current_services=out_region.identity_services, new_service=identity_service
+        )
+        update_region_network_services(
+            current_services=out_region.network_services, new_service=network_service
+        )
+
+    with projects_lock:
+        if project.uuid not in [i.uuid for i in out_projects]:
+            out_projects.append(project)
+
+    with issuer_lock:
+        update_identity_providers(
+            current_issuers=out_issuers, new_issuer=identity_provider
+        )
+
+
+def get_provider(
+    *, os_conf: Provider, trusted_idps: List[Issuer]
+) -> ProviderCreateExtended:
+    """Generate an Openstack virtual provider, reading information from a real openstack
+    instance.
+    """
+    if os_conf.status != ProviderStatus.ACTIVE.value:
+        logger.info(f"Provider={os_conf.name} not active: {os_conf.status}")
+        return ProviderCreateExtended(
+            name=os_conf.name,
+            type=os_conf.type,
+            is_public=os_conf.is_public,
+            support_emails=os_conf.support_emails,
+            status=os_conf.status,
+        )
+
+    trust_idps = copy.deepcopy(trusted_idps)
+    regions: List[RegionCreateExtended] = []
+    projects: List[ProjectCreate] = []
+    identity_providers: List[IdentityProviderCreateExtended] = []
+
+    for region_conf in os_conf.regions:
+        region = RegionCreateExtended(**region_conf.dict())
+        thread_pool = ThreadPoolExecutor(max_workers=len(os_conf.projects))
+        for project_conf in os_conf.projects:
+            thread_pool.submit(
+                get_project_resources,
+                provider_conf=os_conf,
+                project_conf=project_conf,
+                issuers=trust_idps,
+                out_region=region,
+                out_projects=projects,
+                out_issuers=identity_providers,
+            )
+        thread_pool.shutdown(wait=True)
+        regions.append(region)
+
+    # This should be done at the end since we could have skipped some project due to
+    # lack of authorizations or misconfigurations.
+    for region in regions:
+        filter_projects_on_compute_resources(
+            region=region, include_projects=[i.uuid for i in projects]
+        )
+
+    return ProviderCreateExtended(
+        name=os_conf.name,
+        type=os_conf.type,
+        is_public=os_conf.is_public,
+        support_emails=os_conf.support_emails,
+        status=os_conf.status,
+        identity_providers=identity_providers,
+        projects=projects,
+        regions=regions,
+    )
